fix: unable to load a local resource whose path has been encoded
When fetching local static resources, for example, if a local js is introduced into the extension's webviwe, if the path contains spaces (considering that the path in windwos will contain 'Program Files', some apps on mac will have spaces in their names, for example 'Kaitian Ide.app'), it will be encoded as '%20'. This will not be loaded when fetching local resources. Calling decodeURI to handle this.